Split Main in Main and LexGenerator.
This is a minor refactoring to support refacotring  #214
In the future, LexGenerator.generate() will not be static and have internal state in order to support multithreading.
